mesh: narrow MeshState::topology to uint16_t (finding 151)
The classifier bounds the class id to the 16-bit face count and the only
consumer narrows to uint8_t. Cuts the islamic persistent peak by 2532 B
and the hankin peak by 670 B.

+++ core/color/palettes.h
@@ -146,7 +146,8 @@ struct MeshPaletteBank {
    * @param slots Class slot -> palette index (see shuffle_indices).
    * @param out Receives one palette index per face.
    */
-  static HS_COLD_MEMBER void assign_by_class(const int *topology, size_t faces,
+  static HS_COLD_MEMBER void assign_by_class(const uint16_t *topology,
+                                             size_t faces,
                                              const std::array<int, N> &slots,
                                              uint8_t *out) {
     for (size_t f = 0; f < faces; ++f)
